refactor Footer components into a dir
