--- app/components/ui/Footer.scss
@@ -1,128 +1,0 @@
-@import "../../styles/utils/_helpers.scss";
-
-.site-footer {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  flex: 1 0 30vh;
-  color: white;
-  min-height: 100px;
-  background-color: #111928;
-  padding: 96px 0;
-  font-size: 1rem;
-  a {
-    color: white;
-  }
-}
-
-.site-footer__logo {
-  margin-bottom: 2rem;
-}
-
-.site-footer__address {
-  margin-bottom: 1rem;
-  font-weight: 500;
-}
-
-.site-footer__contact {
-  font-size: 14px;
-}
-
-.site-footer__links {
-  flex: 1 1 50%;
-  display: flex;
-  gap: 80px;
-  margin-right: 80px;
-  justify-content: flex-end;
-
-  @media screen and (max-width: 960px) {
-    display: none;
-  }
-
-  figure {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 1rem;
-  }
-
-  figcaption {
-    text-transform: uppercase;
-    font-weight: 600;
-  }
-
-  li {
-    padding-bottom: 1rem;
-    @media screen and (max-width: 960px) {
-      display: none;
-      text-align: center;
-      &:not(:last-child) {
-        padding-right: 0;
-      }
-    }
-  }
-}
-
-.site-footer__content {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  height: 100%;
-  width: 1280px;
-  margin: 0 auto;
-  @media screen and (max-width: $break-tablet-s) {
-    width: auto;
-  }
-}
-
-.site-footer__logos {
-  display: flex;
-  gap: 1rem;
-}
-
-.site-footer__sfseal-center {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-  img {
-    width: 150px;
-    @media screen and (max-width: $break-tablet-s) {
-      width: 120px;
-    }
-  }
-}
-
-.service-guide {
-  flex: 2 2 0;
-  display: flex;
-  justify-content: center;
-  text-align: center;
-  font-weight: 900;
-}
-
-.service-guide__icon {
-  background-image: url(../../assets/img/sfsg-logo.svg);
-  background-repeat: no-repeat;
-  margin-right: 5px;
-  height: 20px;
-  width: 20px;
-}
-
-.service-guide__text a {
-  color: #0025e4;
-  text-transform: uppercase;
-}
-
-.site-footer__top {
-  display: flex;
-}
-
-.site-footer__disclosure {
-  text-align: center;
-  padding-top: 80px;
-  @media screen and (max-width: $break-tablet-s) {
-    display: block;
-    text-align: center;
-    padding: 25px;
-  }
-}
