+++ app/components/ui/Footer/Footer.scss
@@ -1,4 +1,4 @@
-@import "../../styles/utils/_helpers.scss";
+@import "../../../styles/utils/_helpers.scss";
 
 .site-footer {
   display: flex;
@@ -8,39 +8,56 @@
   color: white;
   min-height: 100px;
   background-color: #111928;
-  padding: 50px 0;
-  font-size: 0.9em;
+  padding: 96px 0;
+  font-size: 1rem;
   a {
     color: white;
-    &:hover {
-      color: #0025e4;
-    }
   }
+}
+
+.site-footer__logo {
+  margin-bottom: 2rem;
+}
+
+.site-footer__address {
+  margin-bottom: 1rem;
+  font-weight: 500;
+}
+
+.site-footer__contact {
+  font-size: 14px;
 }
 
 .site-footer__links {
   flex: 1 1 50%;
   display: flex;
-  justify-content: space-evenly;
+  gap: 80px;
+  margin-right: 80px;
+  justify-content: flex-end;
+
   @media screen and (max-width: 960px) {
     display: none;
   }
 
-  & > ul {
-    h1 {
-      font-weight: 900;
-      text-transform: uppercase;
-    }
+  figure {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 1rem;
+  }
 
-    li,
-    h1 {
-      padding-bottom: 10px;
-      @media screen and (max-width: 960px) {
-        display: none;
-        text-align: center;
-        &:not(:last-child) {
-          padding-right: 0;
-        }
+  figcaption {
+    text-transform: uppercase;
+    font-weight: 600;
+  }
+
+  li {
+    padding-bottom: 1rem;
+    @media screen and (max-width: 960px) {
+      display: none;
+      text-align: center;
+      &:not(:last-child) {
+        padding-right: 0;
       }
     }
   }
@@ -56,6 +73,11 @@
   @media screen and (max-width: $break-tablet-s) {
     width: auto;
   }
+}
+
+.site-footer__logos {
+  display: flex;
+  gap: 1rem;
 }
 
 .site-footer__sfseal-center {
@@ -79,7 +101,7 @@
 }
 
 .service-guide__icon {
-  background-image: url(../../assets/img/sfsg-logo.svg);
+  background-image: url(../../../assets/img/sfsg-logo.svg);
   background-repeat: no-repeat;
   margin-right: 5px;
   height: 20px;
