+++ app/components/ui/index.ts
@@ -1,7 +1,7 @@
 export * from "./Banner";
 export { NewsArticles } from "./NewsArticles/NewsArticles";
 export * from "./Datatable";
-export * from "./Footer";
+export * from "./Footer/Footer";
 export * from "./HamburgerMenu";
 export * from "./Loader";
 export * from "./MapElements";
